Use arguments from originalRequest onAuthenticationSuccess
Currently, arguments from the originalRequest are not taken into account onAuthenticationSuccess. So if I requested an object like that (by deeplinking)
http://typo3.flow/my.package/project/edit?project%5B__identity%5D=5b592ade-e810-99b7-179e-75eaabe2d2b5
and editAction is only available for authenticated users, Flow.Login would redirect back to
http://typo3.flow/my.package/project/edit
As a consequence the (mandatory) argument $project would not be set. So I think Flow.Login should respect the arguments of the originalRequest.